+++ pkg/azuremachinepool/validate_azuremachinepool_update.go
@@ -14,19 +14,28 @@ import (
 	expcapzv1alpha3 "sigs.k8s.io/cluster-api-provider-azure/exp/api/v1alpha3"
 	apiv1alpha3 "sigs.k8s.io/cluster-api/api/v1alpha3"
 
+	"github.com/giantswarm/azure-admission-controller/internal/vmcapabilities"
 	"github.com/giantswarm/azure-admission-controller/pkg/validator"
 )
 
 type UpdateValidator struct {
 	k8sClient k8sclient.Interface
 	logger    micrologger.Logger
+	vmcaps    *vmcapabilities.VMSKU
 }
 
 type UpdateValidatorConfig struct {
 	Logger micrologger.Logger
+	VMcaps *vmcapabilities.VMSKU
 }
 
 func NewUpdateValidator(config UpdateValidatorConfig) (*UpdateValidator, error) {
+	if config.Logger == nil {
+		return nil, microerror.Maskf(invalidConfigError, "%T.Logger must not be empty", config)
+	}
+	if config.VMcaps == nil {
+		return nil, microerror.Maskf(invalidConfigError, "%T.VMcaps must not be empty", config)
+	}
 	var k8sClient k8sclient.Interface
 	{
 		restConfig, err := restclient.InClusterConfig()
@@ -53,6 +62,7 @@ func NewUpdateValidator(config UpdateValidatorConfig) (*UpdateValidator, error) 
 	admitter := &UpdateValidator{
 		k8sClient: k8sClient,
 		logger:    config.Logger,
+		vmcaps:    config.VMcaps,
 	}
 
 	return admitter, nil
@@ -68,14 +78,19 @@ func (a *UpdateValidator) Validate(ctx context.Context, request *v1beta1.Admissi
 		return false, microerror.Maskf(parsingFailedError, "unable to parse azureMachinePool CR: %v", err)
 	}
 
-	if azureMPOldCR.Spec.Template.AcceleratedNetworking == nil && azureMPNewCR.Spec.Template.AcceleratedNetworking != nil ||
-		azureMPOldCR.Spec.Template.AcceleratedNetworking != nil && azureMPNewCR.Spec.Template.AcceleratedNetworking == nil {
+	// Instance type might be changed so we check it is still valid and supported.
+	valid, err := checkInstanceTypeIsValid(ctx, a.vmcaps, *azureMPNewCR)
+	if err != nil {
+		return false, microerror.Mask(err)
+	}
+	if !valid {
+		return false, microerror.Maskf(invalidOperationError, "Instance type is invalid or unsupported")
+	}
+
+	if !checkAcceleratedNetworkingUnchanged(ctx, *azureMPOldCR, *azureMPNewCR) {
 		return false, microerror.Maskf(invalidOperationError, "It is not possible to change the AcceleratedNetworking on an existing node pool")
 	}
 
-	if *azureMPOldCR.Spec.Template.AcceleratedNetworking != *azureMPNewCR.Spec.Template.AcceleratedNetworking {
-		return false, microerror.Maskf(invalidOperationError, "It is not possible to change the AcceleratedNetworking on an existing node pool")
-	}
 	return true, nil
 }