+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added validating webhook for AzureNodePools that checks instance type is valid and supports accelerated networking.
+- Added validating webhook for AzureNodePools that checks instance type is valid and meets minimum requirements.
+- Added validating webhook for AzureNodePools that checks instance type supports accelerated networking if enabled.
 
 ## [1.5.0] - 2020-08-19